Guard rm of bundle from script.

+++ submodules-init.sh
@@ -2,16 +2,18 @@
 
 set -e
 
-rm -rf bundle/*
-git rm --cached bundle/*
-rm -rf .git/modules/bundle
+DIR='bundle'
+
+if [ "$(ls -A $DIR)" ]; then
+    rm -rf bundle/*
+    git rm --cached bundle/*
+    rm -rf .git/modules/bundle
+fi
 
 git config -f .gitmodules --get-regexp '^submodule\..*\.path$' |
-    while read path_key path
-    do
-        url_key=$(echo $path_key | sed 's/\.path/.url/')
-        url=$(git config -f .gitmodules --get "$url_key")
-        git submodule add $url $path
-    done
-
-git update-index --assume-unchanged bundle
+while read path_key path
+do
+    url_key=$(echo $path_key | sed 's/\.path/.url/')
+    url=$(git config -f .gitmodules --get "$url_key")
+    git submodule add $url $path
+done
